square with @

+++ src/plugin/sim/components.rs
@@ -1,10 +1,10 @@
 use bevy::prelude::*;
 
-#[derive(Component)]
+#[derive(Component, PartialEq, Eq, Hash)]
 pub struct Position {
-    pub x: u32,
-    pub y: u32,
+    pub x: usize,
+    pub y: usize,
 }
 
-#[derive(Component)]
+#[derive(Component, Clone, Copy)]
 pub struct Symbol(pub char);
